PD#92977 1) add pure loas aac support 2) fix loas AAC PS profile decoder issue

+++ audio_codec/libfaad/mp4.h
@@ -37,12 +37,12 @@ extern "C" {
 
 #include "neaacdec.h"
 
-int8_t AudioSpecificConfig2(uint8_t *pBuffer,
+int AudioSpecificConfig2(uint8_t *pBuffer,
                             uint32_t buffer_size,
                             mp4AudioSpecificConfig *mp4ASC,
                             program_config *pce, uint8_t short_form);
  
-int8_t AudioSpecificConfigFromBitfile(bitfile *ld,
+int AudioSpecificConfigFromBitfile(bitfile *ld,
                                       mp4AudioSpecificConfig *mp4ASC,
                                       program_config *pce, uint32_t bsize, uint8_t short_form);
